+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 3.1.8
 ## New
 - Added counter for Alien Encounters.
+- Added related cards for Starlight Whelp
 
 # 3.1.7
 ## New